docs(api): refactor map and async stuff
